Fixed dra_address_t bug caused by UB
Fixed a bug where the address stored in the sockaddr_storage member of dra_address_t was overwriting the data of contiguous dra_address_t variables because of an oversight in the assumption that all addresses returned by getaddrinfo contain the same address length. The fix now performs a memcpy from the obtained address into the dra_address_t struct and properly stores the address without any problems. This bug would cause programs with multiple dra_address_t members to have undefined behavior, and work only if the variables were defined and edited in a specific order, which is obviously uninteded and clearly an UB mess.

This also meant that the accept function would seem to fail at random because of this problem.

The accept function has been modified to return DRA_NET return codes for consistency, instead of returning the actual socket descriptor as a return value.